// dashboard/src/utils/types/userItem.ts
export interface IUser {
    id?: string;
    firstName: string;
    lastName: string;
    avatar?: string;
    role?: string;
    dob?: string;
  }

  export interface IUserCreate {
    id?: string;
    firstName?: string;
    lastName?: string;
    avatar?: string;
    role?: string;
    dob?: string;
  }
  
  